+++ .gitignore
@@ -2,3 +2,4 @@
 **/.obsidian/**
 **/.DS_Store
 /laboratorio_6cfu/code/bin/**
+**.out
